Move tests from embed to core_api
